requirements.txt
configobj==5.0.6
numpy==1.14.2
opencv-python==3.4.2.17
scipy==1.1.0
setproctitle==1.1.10
tensorflow-gpu==1.3.0
tqdm==4.23.4
